backend/API/requirements.txt
python-dotenv
gunicorn
flask
flask_cors
pymysql
python-jose
boto3
stripe